# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "saitokdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/bookmark/add": {
            "post": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Add a bookmark",
                "operationId": "003d9cab4359d6cd9e0b076a304710f4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Job ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/bookmark/remove": {
            "delete": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Remove a bookmark",
                "operationId": "98cedfac631e1dbce436fe4b62ddf632",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Job ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/bookmark/all/{page}": {
            "get": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Get all categories",
                "operationId": "41ce830bf94379b0cdbd0feea5569362",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/category/add": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add a category",
                "operationId": "b45584f9b122b1e1b592c42cca6df922",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parentId": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/category/get/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get a category",
                "operationId": "cf95d02ca303e2292bf8a40223059473",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/category/all/{root_id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get sub categories",
                "operationId": "ddc1d45431e588b21c3c6eb72f7f1512",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category_root_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/category/delete/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete a category",
                "operationId": "3b785d6f30699297039bc127591771a5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "category_id": {
                "name": "id",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
            "category_root_id": {
                "name": "root_id",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "page": {
                "name": "page",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}